exercise-app/requirements.txt

Flask==1.1.2
python-dotenv==0.16.0
pymongo==3.11.3
certifi==2020.12.5
blinker==1.4
click==7.1.2
Flask-DebugToolbar==0.11.0
flask_login==0.6.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
dnspython==2.7.0
